requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
kaggle
supabase
faker
numpy
fastapi
uvicorn
pydantic
mdpdf
